ensure we can process different types
